refactor: move zip_node to Node runtime
